\name{BIOL6052-package}
\alias{BIOL6052-package}
\alias{BIOL6052}
\keyword{package}

\docType{package}
\title{Supporting material for BIOL6052}
\description{
Datasets and dependencies for BIOL6052 Data Management and Generalised Linear Modelling for Biologists, a 4th year module in the Faculty of Environmental and Life Sciences at the University of Southampton. The datasets are available online, either through R, on Patrick Doncaster's webpages or elsewhere as published material that accompanies the manuscripts. The help files describing the metadata will be discussed within each relevant practical session. 
}
\details{
\tabular{ll}{
Package: \tab BIOL6052\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-01-21\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\authors{
Thomas H.G. Ezard, C. Patrick Doncaster
Maintainer: Thomas H.G. Ezard \email{t.ezard@soton.ac.uk}
}